-  refactor viewModel
