+++ composeApp/src/commonMain/kotlin/com/virtualworld/mipymeanabel/ui/screen/cart/CartViewModel.kt
@@ -1,5 +1,7 @@
 package com.virtualworld.mipymeanabel.ui.screen.cart
 
+import androidx.compose.material3.CalendarLocale
+import androidx.compose.material3.DatePicker
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.virtualworld.mipymeanabel.data.dto.Order
@@ -11,6 +13,8 @@ import com.virtualworld.mipymeanabel.domain.useCase.AuthUseCase
 import com.virtualworld.mipymeanabel.domain.useCase.DeletCartUseCase
 import com.virtualworld.mipymeanabel.domain.useCase.GetProductCartUseCase
 import com.virtualworld.mipymeanabel.ui.screen.model.DataTotals
+import com.virtualworld.mipymeanabel.ui.screen.utils.convertMillisToDate
+import com.virtualworld.mipymeanabel.ui.screen.utils.roundToDecimals
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -18,8 +22,6 @@ import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
-import kotlin.math.pow
-import kotlin.math.round
 
 
 class CartViewModel(
@@ -51,19 +53,7 @@ class CartViewModel(
         checkAuth()
     }
 
-
-    private fun checkAuth() {
-        viewModelScope.launch {
-            authUseCase.loadUser().collect { state ->
-
-                if (state is AuthenticationState.Authenticated) {
-                    _isAuthenticate.value = true
-                }
-
-            }
-        }
-    }
-
+    //Collect products in cart and calculate totals
     private fun getProductsCart() {
         viewModelScope.launch {
             getProductCartUseCase().collect { products ->
@@ -80,12 +70,38 @@ class CartViewModel(
                     }
 
                 }
-                getTotals()
+                calculateTotals()
             }
         }
     }
 
-    private fun getTotals() {
+    //Checking if the user is authenticated
+    private fun checkAuth() {
+        viewModelScope.launch {
+            authUseCase.loadUser().collect { state ->
+
+                if (state is AuthenticationState.Authenticated) {
+                    _isAuthenticate.value = true
+                }
+
+            }
+        }
+    }
+
+    //Update quantity and recalculate
+    fun updateQuantity(idp: Long, unit: Int) {
+
+        _quantity.update { currentQuantity ->
+
+            val newQuantity = currentQuantity[idp]?.plus(unit) ?: 0
+            if (newQuantity >= 0) currentQuantity.plus(idp to newQuantity)
+            else currentQuantity
+        }
+        calculateTotals()
+    }
+
+    //Calculate Totals
+    private fun calculateTotals() {
 
         var totalUSD = 0f
         var totalMN = 0f
@@ -97,44 +113,28 @@ class CartViewModel(
             val priceMN = prod.priceMN
 
             totalUSD += priceUSD * _quantity.value[prod.idp]!!
-
             totalMN += priceMN * _quantity.value[prod.idp]!!
-
             units += _quantity.value[prod.idp]!!
-
         }
 
 
         _totals.update {
-
             DataTotals(
-                totalUSD.roundToDecimals(2), totalMN.roundToDecimals(2), units
+                totalUSD.roundToDecimals(2),
+                totalMN.roundToDecimals(2),
+                units
             )
-
-
         }
     }
 
-    fun updateQuantity(idp: Long, unit: Int) {
-
-        _quantity.update { currentQuantity ->
-
-            val newQuantity = currentQuantity[idp]?.plus(unit) ?: 0
-            if (newQuantity >= 0) currentQuantity.plus(idp to newQuantity)
-            else currentQuantity
-        }
-
-        getTotals()
-
-    }
-
+    //Set Order
     fun onClickAddOrder() {
 
         if (_products.value.isNotEmpty()) {
 
-            val name = "huhuniu"
-            val dateDelviry = "23/223/23"
-            val dateActual = "343/43/43"
+            val name = "Nueva Orden"
+            val dateDelviry = _dateDelivery.toString()
+            val dateActual =  ""  //el repository se encarga
 
             val orderProducts = _products.value.map {
                 OrderProducts(
@@ -157,11 +157,8 @@ class CartViewModel(
 
 
             viewModelScope.launch {
-
-
                 addOrderUseCase(myOrder)
                 deleteCartUseCase.deleteCart()
-
             }
         }
 
@@ -174,39 +171,3 @@ class CartViewModel(
 
 }
 
-fun convertMillisToDate(millis: Long): String {
-    val daysInMonth = intArrayOf(0, 31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31)
-    var days = millis / 86400000 // Milisegundos en un día
-    var year = 1970
-    while (true) {
-        val daysInYear = if (isLeapYear(year)) 366 else 365
-        if (days >= daysInYear) {
-            days -= daysInYear
-            year++
-        } else {
-            break
-        }
-    }
-    var month = 1
-    while (true) {
-        val daysInCurrentMonth = if (month == 2 && isLeapYear(year)) 29 else daysInMonth[month]
-        if (days >= daysInCurrentMonth) {
-            days -= daysInCurrentMonth
-            month++
-        } else {
-            break
-        }
-    }
-    val day = days + 1
-    return "${day.toString().padStart(2, '0')}/${month.toString().padStart(2, '0')}/$year"
-}
-
-fun isLeapYear(year: Int): Boolean {
-    return year % 4 == 0 && (year % 100 != 0 || year % 400 == 0)
-}
-
-
-fun Float.roundToDecimals(decimals: Int): Float {
-    val multiplier = 10.0f.pow(decimals)
-    return round(this * multiplier) / multiplier
-}
